this could solve uidl issue

+++ src/server/core/transaction.h
@@ -8,7 +8,7 @@ typedef struct mailInfo {
     unsigned size;
     bool deleted;
     bool seen;
-    long unsigned int checksum;
+    long long unsigned int checksum;
 } mailInfo;
 
 void transactionOnArrival(const unsigned int state, struct selector_key *key);
